Move ToolConfirmationSubPart into own file
