+++ src/vs/workbench/contrib/chat/browser/chatContentParts/toolInvocationParts/chatToolInvocationPart.ts
@@ -5,45 +5,34 @@
 
 import * as dom from '../../../../../../base/browser/dom.js';
 import { assertNever } from '../../../../../../base/common/assert.js';
-import { RunOnceScheduler } from '../../../../../../base/common/async.js';
 import { decodeBase64 } from '../../../../../../base/common/buffer.js';
 import { Codicon } from '../../../../../../base/common/codicons.js';
 import { Emitter } from '../../../../../../base/common/event.js';
 import { IMarkdownString, MarkdownString } from '../../../../../../base/common/htmlContent.js';
 import { Disposable, DisposableStore, IDisposable, toDisposable } from '../../../../../../base/common/lifecycle.js';
 import { autorunWithStore } from '../../../../../../base/common/observable.js';
-import { count } from '../../../../../../base/common/strings.js';
 import { ThemeIcon } from '../../../../../../base/common/themables.js';
-import { isEmptyObject } from '../../../../../../base/common/types.js';
 import { URI } from '../../../../../../base/common/uri.js';
-import { generateUuid } from '../../../../../../base/common/uuid.js';
 import { MarkdownRenderer } from '../../../../../../editor/browser/widget/markdownRenderer/browser/markdownRenderer.js';
 import { Location } from '../../../../../../editor/common/languages.js';
 import { ILanguageService } from '../../../../../../editor/common/languages/language.js';
 import { IModelService } from '../../../../../../editor/common/services/model.js';
 import { localize } from '../../../../../../nls.js';
-import { ICommandService } from '../../../../../../platform/commands/common/commands.js';
-import { IContextKeyService } from '../../../../../../platform/contextkey/common/contextkey.js';
 import { IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
-import { IKeybindingService } from '../../../../../../platform/keybinding/common/keybinding.js';
-import { IMarkerData, IMarkerService, MarkerSeverity } from '../../../../../../platform/markers/common/markers.js';
-import { ChatContextKeys } from '../../../common/chatContextKeys.js';
 import { IChatMarkdownContent, IChatProgressMessage, IChatTerminalToolInvocationData, IChatToolInvocation, IChatToolInvocationSerialized } from '../../../common/chatService.js';
 import { IChatRendererContent } from '../../../common/chatViewModel.js';
 import { CodeBlockModelCollection } from '../../../common/codeBlockModelCollection.js';
-import { createToolInputUri, createToolSchemaUri, ILanguageModelToolsService, isToolResultInputOutputDetails, IToolResultInputOutputDetails } from '../../../common/languageModelToolsService.js';
-import { CancelChatActionId } from '../../actions/chatExecuteActions.js';
-import { AcceptToolConfirmationActionId } from '../../actions/chatToolActions.js';
-import { ChatTreeItem, IChatCodeBlockInfo, IChatWidgetService } from '../../chat.js';
+import { isToolResultInputOutputDetails, IToolResultInputOutputDetails } from '../../../common/languageModelToolsService.js';
+import { ChatTreeItem, IChatCodeBlockInfo } from '../../chat.js';
 import { getAttachableImageExtension } from '../../chatAttachmentResolve.js';
 import { ICodeBlockRenderOptions } from '../../codeBlockPart.js';
-import { ChatConfirmationWidget, ChatCustomConfirmationWidget, IChatConfirmationButton } from '../chatConfirmationWidget.js';
 import { IChatContentPart, IChatContentPartRenderContext } from '../chatContentParts.js';
 import { ChatMarkdownContentPart, EditorPool } from '../chatMarkdownContentPart.js';
 import { ChatCustomProgressPart, ChatProgressContentPart } from '../chatProgressContentPart.js';
 import { ChatCollapsibleListContentPart, CollapsibleListPool, IChatCollapsibleListItem } from '../chatReferencesContentPart.js';
 import { ChatCollapsibleInputOutputContentPart, IChatCollapsibleIOCodePart, IChatCollapsibleIODataPart } from '../chatToolInputOutputContentPart.js';
 import { TerminalConfirmationWidgetSubPart } from './chatTerminalToolSubPart.js';
+import { ToolConfirmationSubPart } from './chatToolConfirmationSubPart.js';
 
 export class ChatToolInvocationPart extends Disposable implements IChatContentPart {
 	public readonly domNode: HTMLElement;
@@ -63,14 +52,14 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 
 	constructor(
 		private readonly toolInvocation: IChatToolInvocation | IChatToolInvocationSerialized,
-		context: IChatContentPartRenderContext,
-		renderer: MarkdownRenderer,
-		listPool: CollapsibleListPool,
-		editorPool: EditorPool,
-		currentWidthDelegate: () => number,
-		codeBlockModelCollection: CodeBlockModelCollection,
-		codeBlockStartIndex: number,
-		@IInstantiationService instantiationService: IInstantiationService,
+		private readonly context: IChatContentPartRenderContext,
+		private readonly renderer: MarkdownRenderer,
+		private readonly listPool: CollapsibleListPool,
+		private readonly editorPool: EditorPool,
+		private readonly currentWidthDelegate: () => number,
+		private readonly codeBlockModelCollection: CodeBlockModelCollection,
+		private readonly codeBlockStartIndex: number,
+		@IInstantiationService private readonly instantiationService: IInstantiationService,
 	) {
 		super();
 
@@ -87,16 +76,7 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 			dom.clearNode(this.domNode);
 			partStore.clear();
 
-			if (toolInvocation.kind === 'toolInvocation' && toolInvocation.confirmationMessages) {
-				if (toolInvocation.toolSpecificData?.kind === 'terminal') {
-					this.subPart = instantiationService.createInstance(TerminalConfirmationWidgetSubPart, toolInvocation, toolInvocation.toolSpecificData, context, renderer, editorPool, currentWidthDelegate, codeBlockStartIndex);
-				} else {
-					this.subPart = instantiationService.createInstance(ToolConfirmationSubPart, toolInvocation, context, renderer, editorPool, currentWidthDelegate, codeBlockModelCollection, codeBlockStartIndex);
-				}
-			} else {
-				this.subPart = partStore.add(instantiationService.createInstance(ChatToolInvocationSubPart, toolInvocation, context, renderer, listPool, editorPool, currentWidthDelegate, codeBlockModelCollection, codeBlockStartIndex));
-			}
-
+			this.subPart = partStore.add(this.createToolInvocationSubPart());
 			this.domNode.appendChild(this.subPart.domNode);
 			partStore.add(this.subPart.onDidChangeHeight(() => this._onDidChangeHeight.fire()));
 			partStore.add(this.subPart.onNeedsRerender(() => {
@@ -105,6 +85,18 @@ export class ChatToolInvocationPart extends Disposable implements IChatContentPa
 			}));
 		};
 		render();
+	}
+
+	createToolInvocationSubPart(): ChatToolInvocationSubPart | BaseChatToolInvocationSubPart {
+		if (this.toolInvocation.kind === 'toolInvocation' && this.toolInvocation.confirmationMessages) {
+			if (this.toolInvocation.toolSpecificData?.kind === 'terminal') {
+				return this.instantiationService.createInstance(TerminalConfirmationWidgetSubPart, this.toolInvocation, this.toolInvocation.toolSpecificData, this.context, this.renderer, this.editorPool, this.currentWidthDelegate, this.codeBlockStartIndex);
+			} else {
+				return this.instantiationService.createInstance(ToolConfirmationSubPart, this.toolInvocation, this.context, this.renderer, this.editorPool, this.currentWidthDelegate, this.codeBlockModelCollection, this.codeBlockStartIndex);
+			}
+		} else {
+			return this.instantiationService.createInstance(ChatToolInvocationSubPart, this.toolInvocation, this.context, this.renderer, this.listPool, this.editorPool, this.currentWidthDelegate, this.codeBlockModelCollection, this.codeBlockStartIndex);
+		}
 	}
 
 	hasSameContent(other: IChatRendererContent, followingContent: IChatRendererContent[], element: ChatTreeItem): boolean {
@@ -349,257 +341,4 @@ export abstract class BaseChatToolInvocationSubPart extends Disposable {
 	public abstract codeblocks: IChatCodeBlockInfo[];
 
 	public readonly codeblocksPartId = 'tool-' + (BaseChatToolInvocationSubPart.idPool++);
-}
-
-class ToolConfirmationSubPart extends BaseChatToolInvocationSubPart {
-	public readonly domNode: HTMLElement;
-
-	private markdownPart: ChatMarkdownContentPart | undefined;
-	private _codeblocks: IChatCodeBlockInfo[] = [];
-	public get codeblocks(): IChatCodeBlockInfo[] {
-		// TODO this is weird, the separate cases should maybe be their own "subparts"
-		return this.markdownPart?.codeblocks ?? this._codeblocks;
-	}
-
-	constructor(
-		toolInvocation: IChatToolInvocation,
-		private readonly context: IChatContentPartRenderContext,
-		private readonly renderer: MarkdownRenderer,
-		private readonly editorPool: EditorPool,
-		private readonly currentWidthDelegate: () => number,
-		private readonly codeBlockModelCollection: CodeBlockModelCollection,
-		private readonly codeBlockStartIndex: number,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
-		@IKeybindingService keybindingService: IKeybindingService,
-		@IModelService private readonly modelService: IModelService,
-		@ILanguageService private readonly languageService: ILanguageService,
-		@IContextKeyService private readonly contextKeyService: IContextKeyService,
-		@IChatWidgetService private readonly chatWidgetService: IChatWidgetService,
-		@ICommandService private readonly commandService: ICommandService,
-		@IMarkerService private readonly markerService: IMarkerService,
-		@ILanguageModelToolsService private readonly languageModelToolsService: ILanguageModelToolsService,
-	) {
-		super();
-
-		if (!toolInvocation.confirmationMessages) {
-			throw new Error('Confirmation messages are missing');
-		}
-		const { title, message, allowAutoConfirm } = toolInvocation.confirmationMessages;
-		const continueLabel = localize('continue', "Continue");
-		const continueKeybinding = keybindingService.lookupKeybinding(AcceptToolConfirmationActionId)?.getLabel();
-		const continueTooltip = continueKeybinding ? `${continueLabel} (${continueKeybinding})` : continueLabel;
-		const cancelLabel = localize('cancel', "Cancel");
-		const cancelKeybinding = keybindingService.lookupKeybinding(CancelChatActionId)?.getLabel();
-		const cancelTooltip = cancelKeybinding ? `${cancelLabel} (${cancelKeybinding})` : cancelLabel;
-
-		const enum ConfirmationOutcome {
-			Allow,
-			Disallow,
-			AllowWorkspace,
-			AllowGlobally,
-			AllowSession,
-		}
-
-		const buttons: IChatConfirmationButton[] = [
-			{
-				label: continueLabel,
-				data: ConfirmationOutcome.Allow,
-				tooltip: continueTooltip,
-				moreActions: !allowAutoConfirm ? undefined : [
-					{ label: localize('allowSession', 'Allow in this Session'), data: ConfirmationOutcome.AllowSession, tooltip: localize('allowSesssionTooltip', 'Allow this tool to run in this session without confirmation.') },
-					{ label: localize('allowWorkspace', 'Allow in this Workspace'), data: ConfirmationOutcome.AllowWorkspace, tooltip: localize('allowWorkspaceTooltip', 'Allow this tool to run in this workspace without confirmation.') },
-					{ label: localize('allowGlobally', 'Always Allow'), data: ConfirmationOutcome.AllowGlobally, tooltip: localize('allowGloballTooltip', 'Always allow this tool to run without confirmation.') },
-				],
-			},
-			{
-				label: localize('cancel', "Cancel"),
-				data: ConfirmationOutcome.Disallow,
-				isSecondary: true,
-				tooltip: cancelTooltip
-			}];
-		let confirmWidget: ChatConfirmationWidget | ChatCustomConfirmationWidget;
-		if (typeof message === 'string') {
-			confirmWidget = this._register(this.instantiationService.createInstance(
-				ChatConfirmationWidget,
-				title,
-				toolInvocation.originMessage,
-				message,
-				buttons,
-				this.context.container,
-			));
-		} else {
-			const chatMarkdownContent: IChatMarkdownContent = {
-				kind: 'markdownContent',
-				content: message,
-			};
-			const codeBlockRenderOptions: ICodeBlockRenderOptions = {
-				hideToolbar: true,
-				reserveWidth: 19,
-				verticalPadding: 5,
-				editorOptions: {
-					wordWrap: 'on'
-				}
-			};
-
-			const elements = dom.h('div', [
-				dom.h('.message@message'),
-				dom.h('.editor@editor'),
-			]);
-
-			if (toolInvocation.toolSpecificData?.kind === 'input' && toolInvocation.toolSpecificData.rawInput && !isEmptyObject(toolInvocation.toolSpecificData.rawInput)) {
-
-				const inputData = toolInvocation.toolSpecificData;
-
-				const codeBlockRenderOptions: ICodeBlockRenderOptions = {
-					hideToolbar: true,
-					reserveWidth: 19,
-					maxHeightInLines: 13,
-					verticalPadding: 5,
-					editorOptions: {
-						wordWrap: 'off',
-						readOnly: false
-					}
-				};
-
-				const langId = this.languageService.getLanguageIdByLanguageName('json');
-				const rawJsonInput = JSON.stringify(inputData.rawInput ?? {}, null, 1);
-				const canSeeMore = count(rawJsonInput, '\n') > 2; // if more than one key:value
-				const model = this._register(this.modelService.createModel(
-					// View a single JSON line by default until they 'see more'
-					rawJsonInput.replace(/\n */g, ' '),
-					this.languageService.createById(langId),
-					createToolInputUri(toolInvocation.toolId),
-					true
-				));
-
-				const markerOwner = generateUuid();
-				const schemaUri = createToolSchemaUri(toolInvocation.toolId);
-				const validator = new RunOnceScheduler(async () => {
-
-					const newMarker: IMarkerData[] = [];
-
-					const result = await this.commandService.executeCommand('json.validate', schemaUri, model.getValue());
-					for (const item of result) {
-						if (item.range && item.message) {
-							newMarker.push({
-								severity: item.severity === 'Error' ? MarkerSeverity.Error : MarkerSeverity.Warning,
-								message: item.message,
-								startLineNumber: item.range[0].line + 1,
-								startColumn: item.range[0].character + 1,
-								endLineNumber: item.range[1].line + 1,
-								endColumn: item.range[1].character + 1,
-								code: item.code ? String(item.code) : undefined
-							});
-						}
-					}
-
-					this.markerService.changeOne(markerOwner, model.uri, newMarker);
-				}, 500);
-
-				validator.schedule();
-				this._register(model.onDidChangeContent(() => validator.schedule()));
-				this._register(toDisposable(() => this.markerService.remove(markerOwner, [model.uri])));
-				this._register(validator);
-
-				const editor = this._register(this.editorPool.get());
-				editor.object.render({
-					codeBlockIndex: this.codeBlockStartIndex,
-					codeBlockPartIndex: 0,
-					element: this.context.element,
-					languageId: langId ?? 'json',
-					renderOptions: codeBlockRenderOptions,
-					textModel: Promise.resolve(model),
-					chatSessionId: this.context.element.sessionId
-				}, this.currentWidthDelegate());
-				this.codeblocks.push({
-					codeBlockIndex: this.codeBlockStartIndex,
-					codemapperUri: undefined,
-					elementId: this.context.element.id,
-					focus: () => editor.object.focus(),
-					isStreaming: false,
-					ownerMarkdownPartId: this.codeblocksPartId,
-					uri: model.uri,
-					uriPromise: Promise.resolve(model.uri),
-					chatSessionId: this.context.element.sessionId
-				});
-				this._register(editor.object.onDidChangeContentHeight(() => {
-					editor.object.layout(this.currentWidthDelegate());
-					this._onDidChangeHeight.fire();
-				}));
-				this._register(model.onDidChangeContent(e => {
-					try {
-						inputData.rawInput = JSON.parse(model.getValue());
-					} catch {
-						// ignore
-					}
-				}));
-
-				elements.editor.append(editor.object.element);
-
-				if (canSeeMore) {
-					const seeMore = dom.h('div.see-more', [dom.h('a@link')]);
-					seeMore.link.textContent = localize('seeMore', "See more");
-					this._register(dom.addDisposableGenericMouseDownListener(seeMore.link, () => {
-						try {
-							const parsed = JSON.parse(model.getValue());
-							model.setValue(JSON.stringify(parsed, null, 2));
-							editor.object.editor.updateOptions({ wordWrap: 'on' });
-						} catch {
-							// ignored
-						}
-						seeMore.root.remove();
-					}));
-					elements.editor.append(seeMore.root);
-				}
-			}
-
-			this.markdownPart = this._register(this.instantiationService.createInstance(ChatMarkdownContentPart, chatMarkdownContent, this.context, this.editorPool, false, this.codeBlockStartIndex, this.renderer, this.currentWidthDelegate(), this.codeBlockModelCollection, { codeBlockRenderOptions }));
-			elements.message.append(this.markdownPart.domNode);
-
-			this._register(this.markdownPart.onDidChangeHeight(() => this._onDidChangeHeight.fire()));
-			confirmWidget = this._register(this.instantiationService.createInstance(
-				ChatCustomConfirmationWidget,
-				title,
-				toolInvocation.originMessage,
-				elements.root,
-				buttons,
-				this.context.container,
-			));
-		}
-
-		const hasToolConfirmation = ChatContextKeys.Editing.hasToolConfirmation.bindTo(this.contextKeyService);
-		hasToolConfirmation.set(true);
-
-		this._register(confirmWidget.onDidClick(button => {
-			switch (button.data as ConfirmationOutcome) {
-				case ConfirmationOutcome.AllowGlobally:
-					this.languageModelToolsService.setToolAutoConfirmation(toolInvocation.toolId, 'profile', true);
-					toolInvocation.confirmed.complete(true);
-					break;
-				case ConfirmationOutcome.AllowWorkspace:
-					this.languageModelToolsService.setToolAutoConfirmation(toolInvocation.toolId, 'workspace', true);
-					toolInvocation.confirmed.complete(true);
-					break;
-				case ConfirmationOutcome.AllowSession:
-					this.languageModelToolsService.setToolAutoConfirmation(toolInvocation.toolId, 'memory', true);
-					toolInvocation.confirmed.complete(true);
-					break;
-				case ConfirmationOutcome.Allow:
-					toolInvocation.confirmed.complete(true);
-					break;
-				case ConfirmationOutcome.Disallow:
-					toolInvocation.confirmed.complete(false);
-					break;
-			}
-
-			this.chatWidgetService.getWidgetBySessionId(this.context.element.sessionId)?.focusInput();
-		}));
-		this._register(confirmWidget.onDidChangeHeight(() => this._onDidChangeHeight.fire()));
-		this._register(toDisposable(() => hasToolConfirmation.reset()));
-		toolInvocation.confirmed.p.then(() => {
-			hasToolConfirmation.reset();
-			this._onNeedsRerender.fire();
-		});
-		this.domNode = confirmWidget.domNode;
-	}
 }
